\name{hybridEnsembleNews}
\alias{hybridEnsembleNews}
\title{Display the NEWS file}
\usage{
  hybridEnsembleNews()
}
\value{
  None.
}
\description{
  \code{hybridEnsembleNews} shows the NEWS file of the
  hybridEnsemble package.
}
\examples{
hybridEnsembleNews()
}
\author{
  Authors: Michel Ballings and Dirk Van den Poel,
  Maintainer: \email{Michel.Ballings@GMail.com}
}
\references{
  Ballings, M., Vercamer, D., Van den Poel, D., Hybrid
  Ensemble: Many Ensembles is Better Than One, Forthcoming.
}
\seealso{
  \code{\link{hybridEnsemble}},
  \code{\link{predict.hybridEnsemble}},
  \code{\link{importance.hybridEnsemble}},
  \code{\link{CVhybridEnsemble}},
  \code{\link{plot.CVhybridEnsemble}},
  \code{\link{summary.CVhybridEnsemble}}
}

